splitup location_management
